requirements.txt
# File: recall_assistant/requirements.txt

google-cloud-storage
google-cloud-firestore
requests
python-dotenv
llama-parse
openai
